+++ specs/example.spec.ts
@@ -1,5 +1,0 @@
-import { expect, it } from "vitest";
-
-it("should said hello", function () {
-    expect(1).toBe(1);
-});