+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to [Sikessem's UI kit](https://github.com/sikessem/ui) will 
 
 **Full Changelog:** [v0.3.0...v0.4.0](https://github.com/sikessem/ui/compare/v0.4.0...v0.3.0)
 
+- [0.x] Add [color utilities](https://github.com/sikessem/ui#color-utilities)
 - [0.x] Add [icon](https://github.com/sikessem/ui#icon-component) component
 
 ## [v0.4.0](https://github.com/sikessem/ui/releases/tag/v0.4.0) - 2023-07-31